% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_palettes.R
\name{bss_pal}
\alias{bss_pal}
\title{bss_pal}
\usage{
bss_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{define which palette to use, main by default}

\item{reverse}{reverses the palettes' order}

\item{...}{}
}
\value{

}
\description{
bss_pal
}
