% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_palettes.R
\docType{data}
\name{bss_palettes}
\alias{bss_palettes}
\title{define palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
bss_palettes
}
\description{
define palettes
}
\keyword{datasets}
