% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_fill_bss}
\alias{scale_fill_bss}
\title{scale_fill_bss}
\usage{
scale_fill_bss(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
scale_fill_bss
}
