% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss_cols.R
\docType{data}
\name{bss_colors}
\alias{bss_colors}
\title{Define color vectorF and unctionalize the data, this  allows us to get hex colors in a robust and flexible way.}
\format{
An object of class \code{character} of length 5.
}
\usage{
bss_colors
}
\description{
Define color vectorF and unctionalize the data, this  allows us to get hex colors in a robust and flexible way.
}
\keyword{datasets}
