% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_fill_bss}
\alias{scale_fill_bss}
\title{scale_fill_bss}
\usage{
scale_fill_bss(
  palette = "main_only_color",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in bss_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{}
}
\value{

}
\description{
scale_fill_bss
}
